+++ xikolo-ios/ViewControllers/Courses/CourseDecisionViewController.swift
@@ -17,7 +17,7 @@ class CourseDecisionViewController: UIViewController {
         case courseDetails = 2
     }
 
-    @IBOutlet weak var enrollButton: UIBarButtonItem!
+  
     @IBOutlet weak var containerView: UIView!
     @IBOutlet weak var titleView: UILabel!
 
@@ -35,20 +35,11 @@ class CourseDecisionViewController: UIViewController {
         NotificationCenter.default.addObserver(self, selector: #selector(switchViewController), name: NotificationKeys.dropdownCourseContentKey, object: nil)
     }
 
-    @IBAction func enroll(_ sender: UIBarButtonItem) {
-        if UserProfileHelper.isLoggedIn() {
-            showEnrollmentDialog()
-        } else {
-            performSegue(withIdentifier: "ShowLogin", sender: nil)
-        }
-
-    }
-
+  
     @IBAction func unwindSegueToCourseContent(_ segue: UIStoryboardSegue) { }
 
     func decideContent() {
         if(course.enrollment != nil) {
-            navigationItem.rightBarButtonItem = nil
             updateContainerView(course.accessible ? .learnings : .courseDetails)
         } else {
             updateContainerView(.courseDetails)
@@ -102,25 +93,7 @@ class CourseDecisionViewController: UIViewController {
         containerContentViewController = viewController
     }
 
-    func showEnrollmentDialog() {
-        let title = NSLocalizedString("Enroll in course?", comment: "Shown in confirmation dialog")
-        let message = NSLocalizedString("You can always un-enroll.", comment: "Shown in confirmation dialog")
-        let confirm = NSLocalizedString("Enroll", comment: "")
-        let decline = NSLocalizedString("Cancel", comment: "")
-        let alert = UIAlertController(title: title, message: message, preferredStyle: UIAlertControllerStyle.alert)
 
-        alert.addAction(UIAlertAction(title: confirm, style: .default, handler: { (action: UIAlertAction!) in
-            EnrollmentHelper.createEnrollment(for: self.course)
-                .flatMap { CourseHelper.refreshCourses() }
-                .onSuccess { _ in
-                    self.decideContent()
-                }
-        }))
-
-        alert.addAction(UIAlertAction(title: decline, style: .cancel, handler: nil))
-        
-        present(alert, animated: true, completion: nil)
-    }
 
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         switch segue.identifier {
