+++ xikolo-ios/ViewControllers/Courses/CourseDetailViewController.swift
@@ -9,6 +9,7 @@
 import UIKit
 import Hero
 import SDWebImage
+import SimpleRoundedButton
 
 class CourseDetailViewController: UIViewController {
 
@@ -18,13 +19,33 @@ class CourseDetailViewController: UIViewController {
     @IBOutlet weak var dateView: UILabel!
     @IBOutlet weak var teacherView: UILabel!
     @IBOutlet weak var descriptionView: UITextView!
-
+    @IBOutlet weak var enrollmentButton: SimpleRoundedButton!
+    
+    @IBAction func enrollmentButtonClick(_ sender: SimpleRoundedButton) {
+        if UserProfileHelper.isLoggedIn() {
+            showEnrollmentDialog()
+        } else {
+            performSegue(withIdentifier: "ShowLogin", sender: nil)
+        }
+    }
     var course: Course!
     
     override func viewDidLoad() {
         super.viewDidLoad()
 
         titleView.text = course.title
+ 
+        enrollmentButton.setTitle(NSLocalizedString("you_are_enrolled", comment: ""), for: [UIControlState.disabled])
+        enrollmentButton.setTitle(NSLocalizedString("Enroll", comment: ""), for: [UIControlState.normal])
+        if course.enrollment != nil {
+            enrollmentButton.isEnabled = false
+            enrollmentButton.backgroundColor = UIColor.white
+            enrollmentButton.tintColor = Brand.TintColor
+        }else{
+            enrollmentButton.isEnabled = true
+            enrollmentButton.backgroundColor = Brand.TintColor
+            enrollmentButton.tintColor = UIColor.white
+        }
         titleView.heroID = "course_title_" + course.id
         languageView.text = course.language_translated
         languageView.heroID = "course_language_" + course.id
@@ -32,7 +53,7 @@ class CourseDetailViewController: UIViewController {
         teacherView.textColor = Brand.TintColorSecond
         teacherView.heroID = "course_teacher_" + course.id
         imageView.heroID = "course_image_" + course.id
-        
+
         dateView.text = DateLabelHelper.labelFor(startdate: course.start_at, enddate: course.end_at)
         imageView.sd_setImage(with: course.image_url)
 
@@ -41,5 +62,28 @@ class CourseDetailViewController: UIViewController {
             descriptionView.attributedText = markDown
         }
     }
+    
+    func showEnrollmentDialog() {
+        let title = NSLocalizedString("Enroll in course?", comment: "Shown in confirmation dialog")
+        let message = NSLocalizedString("You can always un-enroll.", comment: "Shown in confirmation dialog")
+        let confirm = NSLocalizedString("Enroll", comment: "")
+        let decline = NSLocalizedString("Cancel", comment: "")
+        let alert = UIAlertController(title: title, message: message, preferredStyle: UIAlertControllerStyle.alert)
+        
+        alert.addAction(UIAlertAction(title: confirm, style: .default, handler: { (action: UIAlertAction!) in
+            EnrollmentHelper.createEnrollment(for: self.course)
+                .flatMap { CourseHelper.refreshCourses() }
+                .onSuccess { _ in
+                   //TODO: je anch Kurszustand auf dieser Seite bleiben (neuladen wegen anzeige) oder in Learning
+                
+                    
+            }
+        }))
+        
+        alert.addAction(UIAlertAction(title: decline, style: .cancel, handler: nil))
+        
+        present(alert, animated: true, completion: nil)
+    }
+    
 
 }
